refactor: factorize introduction component in both header
